requirements-dev.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
colorama==0.4.5
crispy-bootstrap5==0.6
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.6
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-storages==1.12.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
gensim==4.2.0
gunicorn==20.1.0
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.29
nbclient==0.6.5
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.22.3
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pywin32==304
pywinpty==2.0.5
pyzmq==23.2.0
report==0.0.1
requests==2.28.1
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
smart-open==6.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
static3==0.7.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.3.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
